added final? methods (need testing on them)
